+++ Sources/KankodaKit/Device/DeviceTypeReader.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2023 Daniel Saidi. All rights reserved.
 //
 
+#if os(macOS) || os(iOS)
 import LocalAuthentication
 import SwiftUI
 
@@ -41,3 +42,4 @@ public extension DeviceTypeReader {
         return type == .touchID
     }
 }
+#endif