+++ Sources/KankodaKit/Items/AppItem/AppItemAuthModifier.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
 //
 
+#if os(macOS) || os(iOS)
 import SwiftUI
 import SwiftUIKit
 
@@ -113,4 +114,5 @@ private extension NotificationCenter {
         publisher(for: NSApplication.didUnhideNotification)
     }
 }
+#endif
 #endif